+++ non_catalog_apps/mousejacker_ms/mousejacker_ducky.c
@@ -87,7 +87,7 @@ static uint32_t mj_ducky_get_command_len(const char* line) {
 
 static bool mj_get_ducky_key(char* key, size_t keylen, MJDuckyKey* dk) {
     //FURI_LOG_D(TAG, "looking up key %s with length %d", key, keylen);
-    for(uint i = 0; i < sizeof(mj_ducky_keys) / sizeof(MJDuckyKey); i++) {
+    for(size_t i = 0; i < sizeof(mj_ducky_keys) / sizeof(MJDuckyKey); i++) {
         if(!strncmp(mj_ducky_keys[i].name, key, keylen)) {
             memcpy(dk, &mj_ducky_keys[i], sizeof(MJDuckyKey));
             return true;
@@ -97,10 +97,10 @@ static bool mj_get_ducky_key(char* key, size_t keylen, MJDuckyKey* dk) {
     return false;
 }
 
-static void checksum(uint8_t* payload, uint len) {
+static void checksum(uint8_t* payload, size_t len) {
     // MS checksum algorithm - as per KeyKeriki paper
     payload[len - 1] = 0x00;
-    for(uint n = 0; n < len - 2; n++) payload[len - 1] ^= payload[n];
+    for(size_t n = 0; n < len - 2; n++) payload[len - 1] ^= payload[n];
     payload[len - 1] = ~payload[len - 1] & 0xff;
 }
